Retargeted for .NET 3.5, removed Common reference
Retargeted .NET 3.5 for portability - works back to XP SP2 and with
Mono.  This required removing the "true" parameter from the BinaryReader
and StreamReaders, but also removing the using block around the readers
results in equivalent operation.

Also removed the reference to QuantumConcepts.Common since the .dll was
built for .NET 4.0, and very little of the functionality was used here.
As a side benefit, the Mono migration analyzer is now happy with
STLdotNET.

Finally, added some unit tests to verify functionality of equality
operators.  The IEnumerable.All method was changed to either .Except or
.Where, which has an added benefit of eliminating facet order from the
STLDocument equality test.