fixed text file detection
in some tested files the IsText function returned true for binary files because the "solid" term text header was present in both text and binary files. The "facet" term (present only in text files on second line) was included for correct text file detection. The tryBinaryIfTextFailed parameter (which could have been a workaround for this) was not used in the original solution.